Multiple updates
Corax: Move camera max speed magic number as game config
Corax: Refactor debug rendering
Corax: Refactor reverse gamepad mapping (input fliped at MovementManager level now)
Corax Add mute decorator to skip the sound business with -mute flag

Corax: Refactor Player: remove useless alternate constructor
Crackle: Add "flip" sprite check
Pluck: Display set animated first frame on the scene viewer
Krita: Fix animation exporter
Script: Add useless script to convert a set to a 3d maya scene (broken)
WCP: Add Rabbit animation going out of the tente
WCP: Add Returns movement/animation during WCP run cycle

+++ sdk/pluck/datas.py
@@ -133,7 +133,7 @@ DATA_TEMPLATES = {
         "type": str,
         "file": str,
         "position": (int, int),
-        "alpha": float,
+        "alpha": int,
         "deph": float
     },
     "player": {
